Add get current account limit endpoint

+++ teacup-loan-accounts/src/main/java/com/example/teacuploanaccounts/entity/AccountRepository.java
@@ -2,5 +2,5 @@ package com.example.teacuploanaccounts.entity;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AccountRepository extends CrudRepository<Account, Long> {
+public interface AccountRepository extends CrudRepository<Account, Integer> {
 }
